refactor: 기존 나의 스페이스 조회 메서드 SpaceMemberRepository에서 SpaceRepository의 getMySpaceList 메서드로 변경
